requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.2
django-celery-results==2.5.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
executing==1.2.0
Faker==18.3.4
idna==3.4
ipython==8.12.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
